Fix some warnings

+++ src/GameWindow.cpp
@@ -75,7 +75,7 @@ void GameWindow::initGui() {
     // need to cope data, because AddFontFromMemoryTTF seems to take control of it, and will try to delete on exit
     char* data = new char[fontData.size()];
     std::memcpy(data, fontData.data(), fontData.size());
-	io.Fonts->AddFontFromMemoryTTF(std::move(data), fontData.size(), 16);
+	io.Fonts->AddFontFromMemoryTTF(std::move(data), static_cast<int>(fontData.size()), 16);
     ImGui_ImplSdlGL3_Init(mWindow);
 
     ImGui::StyleColorsDark();
